AI_module/requirements.txt
fastapi
uvicorn
jwt
PyJWT
python-dotenv
qdrant-client
pydantic
sentence_transformers